se realizo correciones en talleres, AJ, administracion y espacio lugar

+++ app/CamposMagicos/CamposMagicos.php
@@ -22,7 +22,7 @@ class CamposMagicos
     public static function getForeign($table, $campoxxx, $tablaxxx = false)
     {
         $c=CamposMagicos::armarCampo($tablaxxx,$campoxxx);
-        $table->bigInteger($c[1])->unsigned();
+        $table->integer($c[1])->unsigned();
         $table->foreign($c[1])->references('id')->on($c[0]);
         return $table;
     }
@@ -30,7 +30,7 @@ class CamposMagicos
     public static function getForeigDefault($table, $campoxxx, $default,$tablaxxx = false)
     {
         $c=CamposMagicos::armarCampo($tablaxxx,$campoxxx);
-        $table->bigInteger($c[1])->default($default)->unsigned();
+        $table->integer($c[1])->default($default)->unsigned();
         $table->foreign($c[1])->references('id')->on($c[0]);
         return $table;
     }
@@ -43,16 +43,16 @@ class CamposMagicos
     public static function getForeignN($table, $campoxxx, $tablaxxx = false)
     {
         $c=CamposMagicos::armarCampo($tablaxxx,$campoxxx);
-        $table->bigInteger($c[1])->unsigned()->nulleable();
+        $table->integer($c[1])->unsigned()->nulleable();
         $table->foreign($c[1])->references('id')->on($c[0]);
         return $table;
     }
 
     public static function magicos($table)
     {
-        $table->bigInteger('user_crea_id')->unsigned()->default(1)->comment('USUARIO QUE CREAR EL REGISTRO');
-        $table->bigInteger('user_edita_id')->unsigned()->default(1)->comment('USUARIO QUE EDITA EL REGISTRO');
-        $table->bigInteger('sis_esta_id')->unsigned()->default(1)->comment('ESTADO DEL REGISTRO');
+        $table->integer('user_crea_id')->unsigned()->default(1)->comment('USUARIO QUE CREAR EL REGISTRO');
+        $table->integer('user_edita_id')->unsigned()->default(1)->comment('USUARIO QUE EDITA EL REGISTRO');
+        $table->integer('sis_esta_id')->unsigned()->default(1)->comment('ESTADO DEL REGISTRO');
         $table->foreign('sis_esta_id')->references('id')->on('sis_estas');
         $table->foreign('user_crea_id')->references('id')->on('users');
         $table->foreign('user_edita_id')->references('id')->on('users');
@@ -69,7 +69,7 @@ class CamposMagicos
      */
     public static function h_magicos($table)
     {
-        $table->bigInteger('id_old');    // campo nuevo
+        $table->integer('id_old');    // campo nuevo
         $table->integer('user_crea_id');
         $table->integer('user_edita_id');
         $table->integer('sis_esta_id');
