+++ database/migrations/2019_09_04_205096_create_sis_departams_table.php
@@ -16,19 +16,19 @@ class CreateSisDepartamsTable extends Migration
     {
         Schema::create('sis_departams', function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');;
-            $table->bigInteger('sis_pai_id')->unsigned()->nullable()->comment('CAMPO DE ID DEL PAIS');
+            $table->integer('sis_pai_id')->unsigned()->nullable()->comment('CAMPO DE ID DEL PAIS');
             $table->string('s_departamento')->comment('CAMPO DE NOMBRE DEL DEPARTAMENTO');
             $table->Integer('simianti_id')->nullable()->comment('IDENTIFICADOR EN EL SIMI ANTIGUO');
             $table->Integer('user_crea_id');
             $table->integer('user_edita_id');
-            $table->bigInteger('sis_esta_id')->unsigned();
+            $table->integer('sis_esta_id')->unsigned();
             $table->foreign('sis_esta_id')->references('id')->on('sis_estas');
             $table->timestamps();
             $table->foreign('sis_pai_id')->references('id')->on('sis_pais');
         });
         Schema::create('h_sis_departams', function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache();
-            $table->bigInteger('sis_pai_id')->unsigned()->nullable();
+            $table->integer('sis_pai_id')->unsigned()->nullable();
             $table->string('s_departamento');
             $table->Integer('simianti_id')->nullable()->comment('IDENTIFICADOR EN EL SIMI ANTIGUO');
             $table = CamposMagicos::h_magicos($table);
