+++ database/migrations/2019_09_04_205097_create_sis_municipios_table.php
@@ -17,11 +17,11 @@ class CreateSisMunicipiosTable extends Migration
         Schema::create('sis_municipios', function (Blueprint $table) {
 
             $table->increments('id')->start(1)->nocache()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
-            $table->bigInteger('sis_departam_id')->unsigned()->comment('CAMPO DE ID DEL DEPARTAMENTO');
+            $table->integer('sis_departam_id')->unsigned()->comment('CAMPO DE ID DEL DEPARTAMENTO');
             $table->string('s_municipio')->comment('CAMPO NOMBRE DEL MUNICIPIO');
             $table->Integer('user_crea_id');
             $table->integer('user_edita_id');
-            $table->bigInteger('sis_esta_id')->unsigned();
+            $table->integer('sis_esta_id')->unsigned();
             $table->Integer('simianti_id')->nullable()->comment('IDENTIFICADOR EN EL SIMI ANTIGUO');
             $table->foreign('sis_esta_id')->references('id')->on('sis_estas');
             $table->timestamps();
@@ -32,7 +32,7 @@ class CreateSisMunicipiosTable extends Migration
 
             $table->increments('id')->start(1)->nocache();
             $table->Integer('simianti_id')->nullable()->comment('IDENTIFICADOR EN EL SIMI ANTIGUO');
-            $table->bigInteger('sis_departam_id')->unsigned();
+            $table->integer('sis_departam_id')->unsigned();
             $table->string('s_municipio');
             $table = CamposMagicos::h_magicos($table);
         });
