+++ database/migrations/2019_09_04_205937_create_parametros_table.php
@@ -20,7 +20,7 @@ class CreateParametrosTable extends Migration
       $table->string('nombre')->unique()->comment('CAMPO DE NOMBRE DEL PARAMETRO');
       $table->Integer('user_crea_id');
       $table->integer('user_edita_id');
-      $table->bigInteger('sis_esta_id')->unsigned()->default(1);
+      $table->integer('sis_esta_id')->unsigned()->default(1);
       $table->foreign('sis_esta_id')->references('id')->on('sis_estas');
       $table->timestamps();
     });
