+++ database/migrations/2019_09_04_206000_create_sis_menus_table.php
@@ -18,10 +18,10 @@ class CreateSisMenusTable extends Migration
     {
         Schema::create($this->tablaxxx, function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
-            $table->bigInteger('sis_menu_id')->unsigned()->nullable()->comment('CAMPO DEL ID DEL MENU');
+            $table->integer('sis_menu_id')->unsigned()->nullable()->comment('CAMPO DEL ID DEL MENU');
             $table->string('s_menu')->comment('CAMPO NOMBRE DEL MENU');
             $table->string('s_icono')->comment('CAMPO NOMBRE DEL ICONO');
-            $table->bigInteger('sis_docfuen_id')->unsigned()->nullable();
+            $table->integer('sis_docfuen_id')->unsigned()->nullable();
             $table->foreign('sis_menu_id')->references('id')->on('sis_menus');
             $table->foreign('sis_docfuen_id')->references('id')->on('sis_docfuens');
             $table->timestamps();
@@ -33,8 +33,8 @@ class CreateSisMenusTable extends Migration
             $table->increments('id')->start(1)->nocache();
             $table->string('s_menu');
             $table->string('s_icono');
-            $table->bigInteger('sis_docfuen_id')->nullable();
-            $table->bigInteger('sis_menu_id')->nullable();
+            $table->integer('sis_docfuen_id')->nullable();
+            $table->integer('sis_menu_id')->nullable();
             $table = CamposMagicos::h_magicos($table);
         });
        //DB::statement("ALTER TABLE `h_{$this->tablaxxx}` comment 'TABLA QUE ALMACENA LOS LOGS DE LA TABLA h_{$this->tablaxxx}'");
