+++ database/migrations/2019_09_11_172702_create_h_sis_entidads_table.php
@@ -33,8 +33,8 @@ class CreateHSisEntidadsTable extends Migration
        //DB::statement("ALTER TABLE `{$this->tablaxxx2}` comment 'TABLA QUE ALMACENA LOS LOGS DE LA TABLA {$this->tablaxxx2}'");
 
         Schema::create($this->tablaxxx3, function (Blueprint $table) {
-            $table->bigInteger('sis_entidad_id')->unsigned();
-            $table->bigInteger('sis_servicio_id')->unsigned();
+            $table->integer('sis_entidad_id')->unsigned();
+            $table->integer('sis_servicio_id')->unsigned();
             $table = CamposMagicos::h_magicos($table);
         });
        //DB::statement("ALTER TABLE `{$this->tablaxxx3}` comment 'TABLA QUE ALMACENA LOS LOGS DE LA TABLA {$this->tablaxxx3}'");
