+++ database/migrations/2019_09_11_172704_create_sis_procesos_table.php
@@ -17,13 +17,13 @@ class CreateSisProcesosTable extends Migration
     {
         Schema::create($this->tablaxxx, function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
-            $table->bigInteger('sis_proceso_id')->unsigned()->nullable()->comment('CAMPO DE ID DE PROCESO');
-            $table->bigInteger('sis_mapa_proc_id')->unsigned()->comment('CAMPO DE ID MAPA DE PROCESO');
-            $table->bigInteger('prm_proceso_id')->unsigned()->comment('N');
+            $table->integer('sis_proceso_id')->unsigned()->nullable()->comment('CAMPO DE ID DE PROCESO');
+            $table->integer('sis_mapa_proc_id')->unsigned()->comment('CAMPO DE ID MAPA DE PROCESO');
+            $table->integer('prm_proceso_id')->unsigned()->comment('N');
             $table->string('nombre')->comment('CAMPO NOMBRE DEL PROCESO');
-            $table->bigInteger('user_crea_id')->unsigned();
-            $table->bigInteger('user_edita_id')->unsigned();
-            $table->bigInteger('sis_esta_id')->unsigned()->default(1);
+            $table->integer('user_crea_id')->unsigned();
+            $table->integer('user_edita_id')->unsigned();
+            $table->integer('sis_esta_id')->unsigned()->default(1);
             $table->foreign('sis_esta_id')->references('id')->on('sis_estas');
             $table->timestamps();
