+++ database/migrations/2019_09_11_172706_create_h_sis_actividads_table.php
@@ -19,7 +19,7 @@ class CreateHSisActividadsTable extends Migration
         Schema::create($this->tablaxxx, function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache();
             $table->string('nombre');
-            $table->bigInteger('sis_docfuen_id')->unsigned();
+            $table->integer('sis_docfuen_id')->unsigned();
             $table = CamposMagicos::h_magicos($table);
         });
        //DB::statement("ALTER TABLE `{$this->tablaxxx}` comment 'TABLA QUE ALMACENA LOS LOGS DE LA TABLA {$this->tablaxxx}'");
