+++ database/migrations/2019_09_11_173412_create_sis_nnajs_table.php
@@ -18,7 +18,7 @@ class CreateSisNnajsTable extends Migration
     {
         Schema::create($this->tablaxxx, function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
-            $table->bigInteger('prm_escomfam_id')->unsigned()->comment('SABER SI EL REGISTRO QUE ESTA CREANDO ES UN NNAJ O UN COMPONENTE FAMILIAR');
+            $table->integer('prm_escomfam_id')->unsigned()->comment('SABER SI EL REGISTRO QUE ESTA CREANDO ES UN NNAJ O UN COMPONENTE FAMILIAR');
             $table->foreign('prm_escomfam_id')->references('id')->on('parametros');
             $table = CamposMagicos::magicos($table);
         });
