+++ database/migrations/2019_09_16_191649_create_redes_apoyos_table.php
@@ -17,18 +17,18 @@ class CreateRedesApoyosTable extends Migration
   {
     Schema::create($this->tablaxxx, function (Blueprint $table) {
       $table->increments('id')->start(1)->nocache()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
-      $table->bigInteger('entidadAtiende_id')->unsigned()->comment('CAMPO ID DE LA ENTIDAD QUE ATIENDE');
-      $table->bigInteger('ServPrestados_id')->unsigned()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
+      $table->integer('entidadAtiende_id')->unsigned()->comment('CAMPO ID DE LA ENTIDAD QUE ATIENDE');
+      $table->integer('ServPrestados_id')->unsigned()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
       $table->integer('tiempoBeneficio')->comment('CAMPO DE TIEMPO');
-      $table->bigInteger('duracion_id')->unsigned()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
-      $table->bigInteger('tiempoPrestacion_id')->unsigned()->comment('CAMPO DE PARAMETRO DE TIEMPO');
-      $table->bigInteger('tipoRed_id')->unsigned()->comment('CAMPO DE PARAMETRO DE TIPO DE RED');
-      $table->bigInteger('tipoRedPersona_id')->unsigned()->comment('CAMPO DE PARAMETRO TIPO DE PERSONA');
+      $table->integer('duracion_id')->unsigned()->comment('CAMPO DE LLAVE PRIMARIA DE LA TABLA');
+      $table->integer('tiempoPrestacion_id')->unsigned()->comment('CAMPO DE PARAMETRO DE TIEMPO');
+      $table->integer('tipoRed_id')->unsigned()->comment('CAMPO DE PARAMETRO DE TIPO DE RED');
+      $table->integer('tipoRedPersona_id')->unsigned()->comment('CAMPO DE PARAMETRO TIPO DE PERSONA');
       $table->string('nombrePersona')->comment('CAMPO NOMBRE DE LA PERSONA');
-      $table->bigInteger('servBenePersona_id')->unsigned();
-      $table->bigInteger('entidadAtiendeActual_id')->unsigned();
-      $table->bigInteger('durTiempoBen_id')->unsigned();
-      $table->bigInteger('anioPrestServicio_id')->unsigned();
+      $table->integer('servBenePersona_id')->unsigned();
+      $table->integer('entidadAtiendeActual_id')->unsigned();
+      $table->integer('durTiempoBen_id')->unsigned();
+      $table->integer('anioPrestServicio_id')->unsigned();
       $table->string('telApoyo');
       $table->string('dirApoyo');
 
@@ -43,9 +43,9 @@ class CreateRedesApoyosTable extends Migration
       $table->foreign('anioPrestServicio_id')->references('id')->on('parametros');
       $table->foreign('servBenePersona_id')->references('id')->on('parametros');
 
-      $table->bigInteger('user_crea_id')->unsigned();
-      $table->bigInteger('user_edita_id')->unsigned();
-      $table->bigInteger('sis_esta_id')->unsigned();
+      $table->integer('user_crea_id')->unsigned();
+      $table->integer('user_edita_id')->unsigned();
+      $table->integer('sis_esta_id')->unsigned();
       $table->foreign('sis_esta_id')->references('id')->on('sis_estas');
       $table->timestamps();
     });
