+++ database/migrations/2019_09_28_145106_create_h_vsi_antecedentes_table.php
@@ -18,7 +18,7 @@ class CreateHVsiAntecedentesTable extends Migration
     {
         Schema::create($this->tablaxxx, function (Blueprint $table) {
             $table->increments('id')->start(1)->nocache();
-            $table->bigInteger('vsi_id')->unsigned();
+            $table->integer('vsi_id')->unsigned();
             $table->longText('descripcion');
             $table = CamposMagicos::h_magicos($table);
         });
